:sparkles: startBtn.js
Finished the algorithm for the start button and git it working. Imported it to the grid component and prop drilled functtions being called in the startBtn component. I also tested out some of the samples found in the Game of Life's wikipedia page to make sure it was functioning properly